cf/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
google-api-core==2.0.0
google-auth==2.0.1
google-auth-oauthlib==0.4.5
google-cloud-bigquery==2.25.1
google-cloud-bigquery-storage==2.6.3
google-cloud-core==2.0.0
google-crc32c==1.1.2
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
idna==3.2
libcst==0.3.20
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.2
pandas-gbq==0.15.0
proto-plus==1.19.0
protobuf==3.17.3
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
